Changed sort order.
I think this is the best way of sorting example words, it's not perfect
and it doesn't always show many new characters, but I'll stick with
this one. It's very hard to find words that will match all characters
in set, for now this suits me the best. Only way to change that would
be to somehow grade each word and add it to "order by".